+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - feat(MultiSelect): Add `MultiSelect` component
   - fix(input): input only runs validations in onBlur instead of onBlur and onChange
   - fix(number): number only runs validations in onBlur instead of onBlur and onChange
 </details>